Changed some values from int to double. Weights, time and distance for example

+++ TrainingApp/app/src/main/java/com/traininapp/Model/CardioExercise.java
@@ -4,20 +4,20 @@ package com.traininapp.Model;
  * Class CardioExercise is a class for cardio exercises.
  */
 public class CardioExercise extends Exercise{
-    private int runningTime;
-    private int distance;
+    private double runningTime;
+    private double distance;
 
-    public CardioExercise(String name, int runningTime, int distance) {
+    public CardioExercise(String name, double runningTime, double distance) {
         super(name);
         this.runningTime = runningTime;
         this.distance = distance;
     }
 
-    public int getRunningTime() {
+    public double getRunningTime() {
         return runningTime;
     }
 
-    public int getDistance() {
+    public double getDistance() {
         return distance;
     }
 }
